rm codelist and fix accordion
