+++ backend/app/controllers/records_controller.rb
@@ -12,7 +12,7 @@ class RecordsController < ApplicationController
   def show
     render json: @record.to_json(include: [:sources, { entities: { include: [:fields, :properties] } }])
   end
-  
+
 
   def create
     @record = Record.new(record_params)
@@ -28,7 +28,7 @@ class RecordsController < ApplicationController
     # Delete sources that are missing from the request
     source_ids_in_request = params[:record][:sources_attributes].map { |s| s[:id] }
     @record.sources.where.not(id: source_ids_in_request).destroy_all
-    
+
     if @record.update(record_params)
       render json: @record.as_json(include: [:sources, { entities: { include: [:fields, :properties] } }])
     else
@@ -44,22 +44,22 @@ class RecordsController < ApplicationController
 
   def set_record
     @record = Record.includes(:sources, :entities).find(params[:id])
-  end  
+  end
 
   def record_params
     params.require(:record).permit(
       :identifier,
       sources_attributes: [:id, :url, :record_id, :created_at, :updated_at],
       entities_attributes: [
-        :id, 
-        :name, 
+        :id,
+        :name,
         :description,
-        :slug, 
+        :slug,
         :record_id,
-        fields_attributes: [:id, :name, :value], 
+        fields_attributes: [:id, :name, :value],
         properties_attributes: [:id, :name, :value]
       ]
     )
   end
-  
+
 end
